Items and backgrounds split out.
