+++ api_v2/serializers/__init__.py
@@ -1,3 +1,8 @@
 """The initialization for serializers for open5e's api v2."""
 
 from .serializers import *
+
+from .item import ArmorSerializer
+from .item import WeaponSerializer
+from .item import ItemSerializer
+from .item import ItemSetSerializer
